// src/gen/api/types/LikesPreview.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { ReactionUserSummary } from './ReactionUserSummary.ts'

export type LikesPreview = {
  /**
   * @minLength 0
   * @type integer
   */
  count: number
  /**
   * @type array
   */
  preview: ReactionUserSummary[]
}